refactor(native): organize runtime and dx12 subdomain folders
