+++ .specify/specs/005-native-runtime-render-modularization/plan.md
@@ -105,9 +105,9 @@ Not applicable. This feature does not change public Steam copy, store assets, re
 |-------|------|---------------|--------------------|
 | S1 | Finish composition-root decomposition so `engine_composition.c` only wires context and lifecycle entry points | `src/native/engine/runtime/engine_composition.c`, new/existing lifecycle/bootstrap helpers, adjacent tests | `cmake --build out/v3-native` and `ctest --test-dir out/v3-native -C Debug -R "banana_runtime_engine_(host|lifecycle|state)_test|banana_runtime_engine_tick_test|banana_runtime_architecture_dependency_guard_test"` |
 | S2 | Complete tick-flow modularization by extracting any remaining pre-gameplay sequencing into named runtime phases or orchestration helpers without reopening completed input/budget/post seams | `src/native/engine/runtime/engine_tick.c`, new `runtime/tick_*` or `runtime/orchestration/*` helpers, tick-focused tests | `ctest --test-dir out/v3-native -C Debug -R "banana_runtime_engine_tick_test|banana_runtime_tick_input_phase_test|banana_runtime_tick_budget_policy_test|banana_runtime_tick_post_phase_test|banana_engine_input_contract_transition_test"` |
-| S3 | Split DX12 backend diagnostics and transport concerns into explicit backend services with ownership-aligned telemetry and failure surfaces | `src/native/engine/render/backend_dx12.c`, new DX12 helper modules, DX12 tests | `ctest --test-dir out/v3-native -C Debug -R "banana_engine_dx12_runtime_smoke_test|banana_dx12_projection_policy_test|banana_dx12_scene_overlay_frame_test" --output-on-failure` |
+| S3 | Split DX12 backend diagnostics and transport concerns into explicit backend services with ownership-aligned telemetry and failure surfaces | `src/native/engine/render/backend_dx12.c`, new DX12 helper modules under `src/native/engine/render/dx12`, DX12 tests | `ctest --test-dir out/v3-native -C Debug -R "banana_engine_dx12_runtime_smoke_test|banana_dx12_projection_policy_test|banana_dx12_scene_overlay_frame_test" --output-on-failure` |
 | S4 | Shrink the Win32 DX12 POC host so `main.c` becomes a thin runtime shell over objective/scene policies and explicit startup diagnostics | `src/native/engine/win32_dx12_poc/main.c`, POC policy helpers, POC smoke/lane tests | `ctest --test-dir out/v3-native -C Debug -R "banana_engine_win32_dx12_poc_startup_smoke|banana_dx12_objective_policy_test|banana_dx12_scene_flow_test|banana_dx12_gameplay_lane_test|banana_dx12_gameplay_click_lane_test" --output-on-failure` |
-| S5 | Harden governance so concern splits and dependency direction are enforced instead of documented only | `tests/native/runtime_architecture_dependency_guard_test.c`, any supporting guard/checklist artifacts, possibly CMake wiring for new tests | `ctest --test-dir out/v3-native -C Debug -R "banana_runtime_architecture_dependency_guard_test|banana_runtime_engine_tick_test|banana_engine_dx12_runtime_smoke_test" --output-on-failure` |
+| S5 | Harden governance so concern splits and dependency direction are enforced instead of documented only | `tests/native/architecture/runtime_architecture_dependency_guard_test.c`, any supporting guard/checklist artifacts, possibly CMake wiring for new tests | `ctest --test-dir out/v3-native -C Debug -R "banana_runtime_architecture_dependency_guard_test|banana_runtime_engine_tick_test|banana_engine_dx12_runtime_smoke_test" --output-on-failure` |
 
 ## Validation Strategy
