+++ src/native/engine/render/dx12/backend_dx12_diagnostics.c
