+++ src/native/engine/runtime/tick/tick_phases.h
@@ -1,8 +1,8 @@
 #ifndef BANANA_ENGINE_RUNTIME_TICK_PHASES_H
 #define BANANA_ENGINE_RUNTIME_TICK_PHASES_H
 
-#include "../render/renderer.h"
-#include "../render/window.h"
+#include "../../render/renderer.h"
+#include "../../render/window.h"
 
 #ifdef __cplusplus
 extern "C"
