+++ src/native/engine/runtime/tick/tick_post_phase.h
@@ -1,8 +1,8 @@
 #ifndef BANANA_ENGINE_RUNTIME_TICK_POST_PHASE_H
 #define BANANA_ENGINE_RUNTIME_TICK_POST_PHASE_H
 
-#include "../ai/controller.h"
-#include "../world/world.h"
+#include "../../ai/controller.h"
+#include "../../world/world.h"
 
 #ifdef __cplusplus
 extern "C"
