+++ tests/native/architecture/runtime_architecture_dependency_guard_test.c
