+++ tests/native/render/dx12/dx12_projection_policy_test.c
@@ -1,4 +1,4 @@
-#include "render/backend_dx12_projection_policy.h"
+#include "render/dx12/backend_dx12_projection_policy.h"
 
 #include <math.h>
 #include <stdio.h>
