+++ tests/native/runtime/tick/runtime_tick_budget_policy_test.c
@@ -1,4 +1,4 @@
-#include "runtime/tick_budget_policy.h"
+#include "runtime/tick/tick_budget_policy.h"
 
 #include <stdio.h>
