+++ tests/native/runtime/tick/runtime_tick_post_phase_test.c
@@ -1,4 +1,4 @@
-#include "runtime/tick_post_phase.h"
+#include "runtime/tick/tick_post_phase.h"
 
 #include <stdio.h>
